+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - target bucket stage is obtained from `JANIS_ENV` environment variable
+- stage name ENV variable validation
+
 ### Changed
 - target bucket now is a janis-trace-service bucket
 - `add` method now receives client code name instead bucket name