+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.1.0] - 2020-03-04
 ### Added
-- Role assumption for authorizing the Firehose putRecord
+- Optional role assumption for authorizing the Firehose putRecord
 
 ## [3.0.0] - 2020-01-14
 ### Added